% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim-package.R
\docType{package}
\name{clim-package}
\alias{clim}
\alias{clim-package}
\title{clim: 27 Extreme Climate Index}
\description{
27 Extreme Climate Index.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
