% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim.indexes.R
\name{clim.indexes}
\alias{clim.indexes}
\alias{clim.indexes_year}
\alias{index_T}
\alias{index_P}
\title{clim_index}
\usage{
clim.indexes(df)

clim.indexes_year(d, q_ref)

index_T(d, q_ref)

index_P(d, q_ref)
}
\arguments{
\item{q_ref}{quantile of the reference period}
}
\description{
clim_index
}
