% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim-package.R
\name{clim.quantile}
\alias{clim.quantile}
\title{clim.quantile}
\usage{
clim.quantile(df, ref = c(1961, 1990))
}
\value{
quantiles of \code{c("Tmax.10th", "Tmax.90th", "Tmin.10th", "Tmin.90th", "RR.95th", "RR.99th")}
}
\description{
clim.quantile
}
