% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_input.R
\name{tidy_input}
\alias{tidy_input}
\title{tidy_input}
\usage{
tidy_input(data, Segment = "04-03")
}
\arguments{
\item{data}{A data.frame with the columns of \verb{year, month, day, Tavg, Tmax, Tmin, precp}}
}
\description{
tidy_input
}
